requirements/base.txt
django==3.1
djangorestframework==3.11.1
django-extensions==3.0.5
django-filter==2.3.0
